+++ opentelemetry-instrumentation/src/opentelemetry/instrumentation/bootstrap_gen.py
@@ -101,7 +101,11 @@ libraries = [
         "instrumentation": "opentelemetry-instrumentation-jinja2==0.46b0.dev",
     },
     {
-        "library": "kafka-python >= 2.0",
+        "library": "kafka-python >= 2.0, < 3.0",
+        "instrumentation": "opentelemetry-instrumentation-kafka-python==0.46b0.dev",
+    },
+    {
+        "library": "kafka-python-ng >= 2.0, < 3.0",
         "instrumentation": "opentelemetry-instrumentation-kafka-python==0.46b0.dev",
     },
     {